requirements.txt
python-telegram-bot==21.3
python-dotenv==1.0.1
openai==1.35.3
redis==5.0.4
aioredis==2.0.1
httpx==0.27.0
requests==2.32.2
